+++ src/index.ts
@@ -1,0 +1,2 @@
+export { configParseEmoji, parseEmoji } from './parser'
+export { getEmojis } from './emoji'