+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.1.1 (2020-07-01)
+# 2.0.0 (2021-03-19)
 
 
 ### Bug Fixes
@@ -9,12 +9,32 @@
 
 ### Features
 
+* add 2.x 解析功能 ([2a7bd94](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/2a7bd94b5a382ed74e8d2310e4be6d651ba7f27e))
+* add 2.x 解析功能 ([fd344a1](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/fd344a1829ad18c3a327c5a64b90b729a76ae078))
 * add emoji chars ([72f3fb0](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/72f3fb00131af676493949ba4fb30ddafde48077))
 * add emoji panel ([464a9aa](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/464a9aac666070c2816bdac9ac68406dad74f40a))
 * add emoji parser ([e22ec89](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/e22ec893a73c4eee85f0e8bf6ea40827c42b149c))
+* add example for vue ([3c2f777](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/3c2f77781314b1f940595b8f2bc9335b5875f50f))
+* add example for vue ([dc4f30f](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/dc4f30f39be14a9f653461888092353dc277cf7d))
+* add getEmojis ([092d6c5](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/092d6c517c387ad321de288aa5b262aa83cb7944))
 * add more emoji ([14a409f](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/14a409f488cbc4eee29f970e7812ec4fe0eaef44))
 * add structor ([9e2e425](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/9e2e42533dc9017611d57fc215eb245b25332341))
 * 添加更多表情 ([f16451c](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/f16451c540a58abe21815d639e59e43834204add))
 
 
 
+## 1.1.1 (2020-07-01)
+
+### Bug Fixes
+
+- add dist dir ([1d60ee3](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/1d60ee36c857a46833969852da9584c2200bdf23))
+- multiple emoji error ([653aa7f](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/653aa7f0921d4f091584f174e822cd680004f2af))
+
+### Features
+
+- add emoji chars ([72f3fb0](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/72f3fb00131af676493949ba4fb30ddafde48077))
+- add emoji panel ([464a9aa](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/464a9aac666070c2816bdac9ac68406dad74f40a))
+- add emoji parser ([e22ec89](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/e22ec893a73c4eee85f0e8bf6ea40827c42b149c))
+- add more emoji ([14a409f](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/14a409f488cbc4eee29f970e7812ec4fe0eaef44))
+- add structor ([9e2e425](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/9e2e42533dc9017611d57fc215eb245b25332341))
+- 添加更多表情 ([f16451c](https://github.com/mingtianyihou33/wechat-emoji-parser/commit/f16451c540a58abe21815d639e59e43834204add))